test: Improve setup
